# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "the api to query the information about *Star War*\nyou can check all\
    \ the Star Wars data you've ever wanted\nPlanets Spaceships Vehicles People Films\
    \ and Species\nFrom all SEVEN Star Wars films\n"
  version: "0.0.0"
  title: "star world API"
paths:
  /planets:
    get:
      summary: "get planet list"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page of the list queried\n"
        required: false
        type: "integer"
        default: 0
        minimum: 0
        x-exportParamName: "Page"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/planets_list"
        404:
          description: "资源不存在\n"
  /starships:
    get:
      summary: "get starships list"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page of the list queried\n"
        required: false
        type: "integer"
        default: 0
        minimum: 0
        x-exportParamName: "Page"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/starships_list"
        404:
          description: "资源不存在\n"
  /vehicles:
    get:
      summary: "get vehicles list"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page of the list queried\n"
        required: false
        type: "integer"
        default: 0
        minimum: 0
        x-exportParamName: "Page"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/vehicles_list"
        404:
          description: "资源不存在\n"
  /people:
    get:
      summary: "get people list"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page of the list queried\n"
        required: false
        type: "integer"
        default: 0
        minimum: 0
        x-exportParamName: "Page"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/people_list"
        404:
          description: "资源不存在\n"
  /films:
    get:
      summary: "get films list"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page of the list queried\n"
        required: false
        type: "integer"
        default: 0
        minimum: 0
        x-exportParamName: "Page"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/films_list"
        404:
          description: "资源不存在\n"
  /species:
    get:
      summary: "get species list"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Page of the list queried\n"
        required: false
        type: "integer"
        default: 0
        minimum: 0
        x-exportParamName: "Page"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/people_list"
        404:
          description: "资源不存在\n"
...
definitions:
...
  planets_list:
    allOf:
    - $ref: "#/definitions/link_list"
    - type: "object"
      properties:
        results:
          type: "array"
          items:
            $ref: "#/definitions/planets"
    example:
      count: 61
      next: "https://swapi.co/api/planets/?page=2"
      previous: null
      results:
      - name: "Alderaan"
        rotation_period: "24"
        orbital_period: "364"
        diameter: "12500"
        climate: "temperate"
        gravity: "1 standard"
        terrain: "grasslands, mountains"
        surface_water: "40"
        population: "2000000000"
        residents:
        - "https://swapi.co/api/people/5/"
        - "https://swapi.co/api/people/68/"
        - "https://swapi.co/api/people/81/"
        films:
        - "https://swapi.co/api/films/6/"
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T11:35:48.479000Z"
        edited: "2014-12-20T20:58:18.420000Z"
        url: "https://swapi.co/api/planets/2/"
      - name: "Yavin IV"
        rotation_period: "24"
        orbital_period: "4818"
        diameter: "10200"
        climate: "temperate, tropical"
        gravity: "1 standard"
        terrain: "jungle, rainforests"
        surface_water: "8"
        population: "1000"
        residents: []
        films:
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T11:37:19.144000Z"
        edited: "2014-12-20T20:58:18.421000Z"
        url: "https://swapi.co/api/planets/3/"
  starships_list:
    allOf:
    - $ref: "#/definitions/link_list"
    - type: "object"
      properties:
        results:
          type: "array"
          items:
            $ref: "#/definitions/starships"
    example:
      count: 37
      next: "https://swapi.co/api/starships/?page=2"
      previous: null
      results:
      - name: "Executor"
        model: "Executor-class star dreadnought"
        manufacturer: "Kuat Drive Yards, Fondor Shipyards"
        cost_in_credits: "1143350000"
        length: "19000"
        max_atmosphering_speed: "n/a"
        crew: "279144"
        passengers: "38000"
        cargo_capacity: "250000000"
        consumables: "6 years"
        hyperdrive_rating: "2.0"
        MGLT: "40"
        starship_class: "Star dreadnought"
        pilots: []
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        created: "2014-12-15T12:31:42.547000Z"
        edited: "2017-04-19T10:56:06.685592Z"
        url: "https://swapi.co/api/starships/15/"
      - name: "Sentinel-class landing craft"
        model: "Sentinel-class landing craft"
        manufacturer: "Sienar Fleet Systems, Cyngus Spaceworks"
        cost_in_credits: "240000"
        length: "38"
        max_atmosphering_speed: "1000"
        crew: "5"
        passengers: "75"
        cargo_capacity: "180000"
        consumables: "1 month"
        hyperdrive_rating: "1.0"
        MGLT: "70"
        starship_class: "landing craft"
        pilots: []
        films:
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T15:48:00.586000Z"
        edited: "2014-12-22T17:35:44.431407Z"
        url: "https://swapi.co/api/starships/5/"
      - name: "Death Star"
        model: "DS-1 Orbital Battle Station"
        manufacturer: "Imperial Department of Military Research, Sienar Fleet Systems"
        cost_in_credits: "1000000000000"
        length: "120000"
        max_atmosphering_speed: "n/a"
        crew: "342953"
        passengers: "843342"
        cargo_capacity: "1000000000000"
        consumables: "3 years"
        hyperdrive_rating: "4.0"
        MGLT: "10"
        starship_class: "Deep Space Mobile Battlestation"
        pilots: []
        films:
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T16:36:50.509000Z"
        edited: "2014-12-22T17:35:44.452589Z"
        url: "https://swapi.co/api/starships/9/"
      - name: "Millennium Falcon"
        model: "YT-1300 light freighter"
        manufacturer: "Corellian Engineering Corporation"
        cost_in_credits: "100000"
        length: "34.37"
        max_atmosphering_speed: "1050"
        crew: "4"
        passengers: "6"
        cargo_capacity: "100000"
        consumables: "2 months"
        hyperdrive_rating: "0.5"
        MGLT: "75"
        starship_class: "Light freighter"
        pilots:
        - "https://swapi.co/api/people/13/"
        - "https://swapi.co/api/people/14/"
        - "https://swapi.co/api/people/25/"
        - "https://swapi.co/api/people/31/"
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/7/"
        - "https://swapi.co/api/films/3/"
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T16:59:45.094000Z"
        edited: "2014-12-22T17:35:44.464156Z"
        url: "https://swapi.co/api/starships/10/"
      - name: "Y-wing"
        model: "BTL Y-wing"
        manufacturer: "Koensayr Manufacturing"
        cost_in_credits: "134999"
        length: "14"
        max_atmosphering_speed: "1000km"
        crew: "2"
        passengers: "0"
        cargo_capacity: "110"
        consumables: "1 week"
        hyperdrive_rating: "1.0"
        MGLT: "80"
        starship_class: "assault starfighter"
        pilots: []
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        - "https://swapi.co/api/films/1/"
        created: "2014-12-12T11:00:39.817000Z"
        edited: "2014-12-22T17:35:44.479706Z"
        url: "https://swapi.co/api/starships/11/"
      - name: "X-wing"
        model: "T-65 X-wing"
        manufacturer: "Incom Corporation"
        cost_in_credits: "149999"
        length: "12.5"
        max_atmosphering_speed: "1050"
        crew: "1"
        passengers: "0"
        cargo_capacity: "110"
        consumables: "1 week"
        hyperdrive_rating: "1.0"
        MGLT: "100"
        starship_class: "Starfighter"
        pilots:
        - "https://swapi.co/api/people/1/"
        - "https://swapi.co/api/people/9/"
        - "https://swapi.co/api/people/18/"
        - "https://swapi.co/api/people/19/"
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        - "https://swapi.co/api/films/1/"
        created: "2014-12-12T11:19:05.340000Z"
        edited: "2014-12-22T17:35:44.491233Z"
        url: "https://swapi.co/api/starships/12/"
      - name: "TIE Advanced x1"
        model: "Twin Ion Engine Advanced x1"
        manufacturer: "Sienar Fleet Systems"
        cost_in_credits: "unknown"
        length: "9.2"
        max_atmosphering_speed: "1200"
        crew: "1"
        passengers: "0"
        cargo_capacity: "150"
        consumables: "5 days"
        hyperdrive_rating: "1.0"
        MGLT: "105"
        starship_class: "Starfighter"
        pilots:
        - "https://swapi.co/api/people/4/"
        films:
        - "https://swapi.co/api/films/1/"
        created: "2014-12-12T11:21:32.991000Z"
        edited: "2014-12-22T17:35:44.549047Z"
        url: "https://swapi.co/api/starships/13/"
      - name: "Slave 1"
        model: "Firespray-31-class patrol and attack"
        manufacturer: "Kuat Systems Engineering"
        cost_in_credits: "unknown"
        length: "21.5"
        max_atmosphering_speed: "1000"
        crew: "1"
        passengers: "6"
        cargo_capacity: "70000"
        consumables: "1 month"
        hyperdrive_rating: "3.0"
        MGLT: "70"
        starship_class: "Patrol craft"
        pilots:
        - "https://swapi.co/api/people/22/"
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/5/"
        created: "2014-12-15T13:00:56.332000Z"
        edited: "2014-12-22T17:35:44.716273Z"
        url: "https://swapi.co/api/starships/21/"
      - name: "Imperial shuttle"
        model: "Lambda-class T-4a shuttle"
        manufacturer: "Sienar Fleet Systems"
        cost_in_credits: "240000"
        length: "20"
        max_atmosphering_speed: "850"
        crew: "6"
        passengers: "20"
        cargo_capacity: "80000"
        consumables: "2 months"
        hyperdrive_rating: "1.0"
        MGLT: "50"
        starship_class: "Armed government transport"
        pilots:
        - "https://swapi.co/api/people/1/"
        - "https://swapi.co/api/people/13/"
        - "https://swapi.co/api/people/14/"
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        created: "2014-12-15T13:04:47.235000Z"
        edited: "2014-12-22T17:35:44.795405Z"
        url: "https://swapi.co/api/starships/22/"
      - name: "EF76 Nebulon-B escort frigate"
        model: "EF76 Nebulon-B escort frigate"
        manufacturer: "Kuat Drive Yards"
        cost_in_credits: "8500000"
        length: "300"
        max_atmosphering_speed: "800"
        crew: "854"
        passengers: "75"
        cargo_capacity: "6000000"
        consumables: "2 years"
        hyperdrive_rating: "2.0"
        MGLT: "40"
        starship_class: "Escort ship"
        pilots: []
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        created: "2014-12-15T13:06:30.813000Z"
        edited: "2014-12-22T17:35:44.848329Z"
        url: "https://swapi.co/api/starships/23/"
  vehicles_list:
    allOf:
    - $ref: "#/definitions/link_list"
    - type: "object"
      properties:
        results:
          type: "array"
          items:
            $ref: "#/definitions/vehicles"
    example:
      count: 39
      next: "https://swapi.co/api/vehicles/?page=2"
      previous: null
      results:
      - name: "Sand Crawler"
        model: "Digger Crawler"
        manufacturer: "Corellia Mining Corporation"
        cost_in_credits: "150000"
        length: "36.8"
        max_atmosphering_speed: "30"
        crew: "46"
        passengers: "30"
        cargo_capacity: "50000"
        consumables: "2 months"
        vehicle_class: "wheeled"
        pilots: []
        films:
        - "https://swapi.co/api/films/5/"
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T15:36:25.724000Z"
        edited: "2014-12-22T18:21:15.523587Z"
        url: "https://swapi.co/api/vehicles/4/"
      - name: "T-16 skyhopper"
        model: "T-16 skyhopper"
        manufacturer: "Incom Corporation"
        cost_in_credits: "14500"
        length: "10.4"
        max_atmosphering_speed: "1200"
        crew: "1"
        passengers: "1"
        cargo_capacity: "50"
        consumables: "0"
        vehicle_class: "repulsorcraft"
        pilots: []
        films:
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T16:01:52.434000Z"
        edited: "2014-12-22T18:21:15.552614Z"
        url: "https://swapi.co/api/vehicles/6/"
      - name: "X-34 landspeeder"
        model: "X-34 landspeeder"
        manufacturer: "SoroSuub Corporation"
        cost_in_credits: "10550"
        length: "3.4"
        max_atmosphering_speed: "250"
        crew: "1"
        passengers: "1"
        cargo_capacity: "5"
        consumables: "unknown"
        vehicle_class: "repulsorcraft"
        pilots: []
        films:
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T16:13:52.586000Z"
        edited: "2014-12-22T18:21:15.583700Z"
        url: "https://swapi.co/api/vehicles/7/"
      - name: "TIE/LN starfighter"
        model: "Twin Ion Engine/Ln Starfighter"
        manufacturer: "Sienar Fleet Systems"
        cost_in_credits: "unknown"
        length: "6.4"
        max_atmosphering_speed: "1200"
        crew: "1"
        passengers: "0"
        cargo_capacity: "65"
        consumables: "2 days"
        vehicle_class: "starfighter"
        pilots: []
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        - "https://swapi.co/api/films/1/"
        created: "2014-12-10T16:33:52.860000Z"
        edited: "2014-12-22T18:21:15.606149Z"
        url: "https://swapi.co/api/vehicles/8/"
      - name: "Snowspeeder"
        model: "t-47 airspeeder"
        manufacturer: "Incom corporation"
        cost_in_credits: "unknown"
        length: "4.5"
        max_atmosphering_speed: "650"
        crew: "2"
        passengers: "0"
        cargo_capacity: "10"
        consumables: "none"
        vehicle_class: "airspeeder"
        pilots:
        - "https://swapi.co/api/people/1/"
        - "https://swapi.co/api/people/18/"
        films:
        - "https://swapi.co/api/films/2/"
        created: "2014-12-15T12:22:12Z"
        edited: "2014-12-22T18:21:15.623033Z"
        url: "https://swapi.co/api/vehicles/14/"
      - name: "TIE bomber"
        model: "TIE/sa bomber"
        manufacturer: "Sienar Fleet Systems"
        cost_in_credits: "unknown"
        length: "7.8"
        max_atmosphering_speed: "850"
        crew: "1"
        passengers: "0"
        cargo_capacity: "none"
        consumables: "2 days"
        vehicle_class: "space/planetary bomber"
        pilots: []
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        created: "2014-12-15T12:33:15.838000Z"
        edited: "2014-12-22T18:21:15.667730Z"
        url: "https://swapi.co/api/vehicles/16/"
      - name: "AT-AT"
        model: "All Terrain Armored Transport"
        manufacturer: "Kuat Drive Yards, Imperial Department of Military Research"
        cost_in_credits: "unknown"
        length: "20"
        max_atmosphering_speed: "60"
        crew: "5"
        passengers: "40"
        cargo_capacity: "1000"
        consumables: "unknown"
        vehicle_class: "assault walker"
        pilots: []
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        created: "2014-12-15T12:38:25.937000Z"
        edited: "2014-12-22T18:21:15.714673Z"
        url: "https://swapi.co/api/vehicles/18/"
      - name: "AT-ST"
        model: "All Terrain Scout Transport"
        manufacturer: "Kuat Drive Yards, Imperial Department of Military Research"
        cost_in_credits: "unknown"
        length: "2"
        max_atmosphering_speed: "90"
        crew: "2"
        passengers: "0"
        cargo_capacity: "200"
        consumables: "none"
        vehicle_class: "walker"
        pilots:
        - "https://swapi.co/api/people/13/"
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/3/"
        created: "2014-12-15T12:46:42.384000Z"
        edited: "2014-12-22T18:21:15.761584Z"
        url: "https://swapi.co/api/vehicles/19/"
      - name: "Storm IV Twin-Pod cloud car"
        model: "Storm IV Twin-Pod"
        manufacturer: "Bespin Motors"
        cost_in_credits: "75000"
        length: "7"
        max_atmosphering_speed: "1500"
        crew: "2"
        passengers: "0"
        cargo_capacity: "10"
        consumables: "1 day"
        vehicle_class: "repulsorcraft"
        pilots: []
        films:
        - "https://swapi.co/api/films/2/"
        created: "2014-12-15T12:58:50.530000Z"
        edited: "2014-12-22T18:21:15.783232Z"
        url: "https://swapi.co/api/vehicles/20/"
      - name: "Sail barge"
        model: "Modified Luxury Sail Barge"
        manufacturer: "Ubrikkian Industries Custom Vehicle Division"
        cost_in_credits: "285000"
        length: "30"
        max_atmosphering_speed: "100"
        crew: "26"
        passengers: "500"
        cargo_capacity: "2000000"
        consumables: "Live food tanks"
        vehicle_class: "sail barge"
        pilots: []
        films:
        - "https://swapi.co/api/films/3/"
        created: "2014-12-18T10:44:14.217000Z"
        edited: "2014-12-22T18:21:15.807906Z"
        url: "https://swapi.co/api/vehicles/24/"
  people_list:
    allOf:
    - $ref: "#/definitions/link_list"
    - type: "object"
      properties:
        results:
          type: "array"
          items:
            $ref: "#/definitions/people"
    example:
      count: 87
      next: "https://swapi.co/api/people/?page=2"
      previous: null
      results:
      - name: "Luke Skywalker"
        height: "172"
        mass: "77"
        hair_color: "blond"
        skin_color: "fair"
        eye_color: "blue"
        birth_year: "19BBY"
        gender: "male"
        homeworld: "https://swapi.co/api/planets/1/"
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/6/"
        - "https://swapi.co/api/films/3/"
        - "https://swapi.co/api/films/1/"
        - "https://swapi.co/api/films/7/"
        species:
        - "https://swapi.co/api/species/1/"
        vehicles:
        - "https://swapi.co/api/vehicles/14/"
        - "https://swapi.co/api/vehicles/30/"
        starships:
        - "https://swapi.co/api/starships/12/"
        - "https://swapi.co/api/starships/22/"
        created: "2014-12-09T13:50:51.644000Z"
        edited: "2014-12-20T21:17:56.891000Z"
        url: "https://swapi.co/api/people/1/"
      - name: "C-3PO"
        height: "167"
        mass: "75"
        hair_color: "n/a"
        skin_color: "gold"
        eye_color: "yellow"
        birth_year: "112BBY"
        gender: "n/a"
        homeworld: "https://swapi.co/api/planets/1/"
        films:
        - "https://swapi.co/api/films/2/"
        - "https://swapi.co/api/films/5/"
        - "https://swapi.co/api/films/4/"
        - "https://swapi.co/api/films/6/"
        - "https://swapi.co/api/films/3/"
        - "https://swapi.co/api/films/1/"
        species:
        - "https://swapi.co/api/species/2/"
        vehicles: []
        starships: []
        created: "2014-12-10T15:10:51.357000Z"
        edited: "2014-12-20T21:17:50.309000Z"
        url: "https://swapi.co/api/people/2/"
  films_list:
    allOf:
    - $ref: "#/definitions/link_list"
    - type: "object"
      properties:
        results:
          type: "array"
          items:
            $ref: "#/definitions/films"
    example:
      count: 7
      next: null
      previous: null
      results:
      - title: "A New Hope"
        episode_id: 4
        opening_crawl: "It is a period of civil war.\r\nRebel spaceships, striking\r\
          \nfrom a hidden base, have won\r\ntheir first victory against\r\nthe evil\
          \ Galactic Empire.\r\n\r\nDuring the battle, Rebel\r\nspies managed to steal\
          \ secret\r\nplans to the Empire's\r\nultimate weapon, the DEATH\r\nSTAR,\
          \ an armored space\r\nstation with enough power\r\nto destroy an entire\
          \ planet.\r\n\r\nPursued by the Empire's\r\nsinister agents, Princess\r\n\
          Leia races home aboard her\r\nstarship, custodian of the\r\nstolen plans\
          \ that can save her\r\npeople and restore\r\nfreedom to the galaxy...."
        director: "George Lucas"
        producer: "Gary Kurtz, Rick McCallum"
        release_date: "1977-05-25"
        characters:
        - "https://swapi.co/api/people/1/"
        - "https://swapi.co/api/people/2/"
        - "https://swapi.co/api/people/3/"
        - "https://swapi.co/api/people/4/"
        - "https://swapi.co/api/people/5/"
        - "https://swapi.co/api/people/6/"
        - "https://swapi.co/api/people/7/"
        - "https://swapi.co/api/people/8/"
        - "https://swapi.co/api/people/9/"
        - "https://swapi.co/api/people/10/"
        - "https://swapi.co/api/people/12/"
        - "https://swapi.co/api/people/13/"
        - "https://swapi.co/api/people/14/"
        - "https://swapi.co/api/people/15/"
        - "https://swapi.co/api/people/16/"
        - "https://swapi.co/api/people/18/"
        - "https://swapi.co/api/people/19/"
        - "https://swapi.co/api/people/81/"
        planets:
        - "https://swapi.co/api/planets/2/"
        - "https://swapi.co/api/planets/3/"
        - "https://swapi.co/api/planets/1/"
        starships:
        - "https://swapi.co/api/starships/2/"
        - "https://swapi.co/api/starships/3/"
        - "https://swapi.co/api/starships/5/"
        - "https://swapi.co/api/starships/9/"
        - "https://swapi.co/api/starships/10/"
        - "https://swapi.co/api/starships/11/"
        - "https://swapi.co/api/starships/12/"
        - "https://swapi.co/api/starships/13/"
        vehicles:
        - "https://swapi.co/api/vehicles/4/"
        - "https://swapi.co/api/vehicles/6/"
        - "https://swapi.co/api/vehicles/7/"
        - "https://swapi.co/api/vehicles/8/"
        species:
        - "https://swapi.co/api/species/5/"
        - "https://swapi.co/api/species/3/"
        - "https://swapi.co/api/species/2/"
        - "https://swapi.co/api/species/1/"
        - "https://swapi.co/api/species/4/"
        created: "2014-12-10T14:23:31.880000Z"
        edited: "2015-04-11T09:46:52.774897Z"
        url: "https://swapi.co/api/films/1/"
...